Investigate Firefox navigation
I've seen Firefox navigate away from the test page to a Yahoo page, searching for a thang type url.
Testing this by reverting to having Travis run Firefox, enabling the previously disabled tests,
and adding logging for when thang portrait and user avatar urls are generated, and when the page
is navigated away from.